requirements.txt
altair==4.2.2
altgraph==0.17.3
ansunit==0.1.7
anyio==3.7.1
Arpeggio==2.0.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
black==23.7.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
clingo==5.6.2
clyngor==0.4.3
colorama==0.4.6
contourpy==1.1.0
coverage==7.2.7
cycler==0.11.0
decorator==5.1.1
draco==2.0.0b5
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.95.2
fonttools==4.41.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mypy==1.4.1
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.2.0
prompt-toolkit==3.0.39
protobuf==4.23.4
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.0
Pygments==2.15.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.3
Pympler==1.0.1
PyNLPIR==0.6.0
pyparsing==3.0.9
pyPEG2==2.15.2
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==5.4.1
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.21.1
validators==0.20.0
vega-datasets==0.9.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.16.2
